+++ src/app/app.component.ts
@@ -249,15 +249,13 @@ export class AppComponent {
 
     // ** NOTIFICATIONS - Anchor Status **
     this.obsList.push(
-      this.alarmsFacade.anchorStatus$().subscribe((r) => {
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      this.alarmsFacade.anchorStatus$().subscribe((r: any) => {
         if (r.error) {
-          if (r.result === 401) {
+          if (r.result.status === 401) {
             this.showLogin();
           } else {
-            this.app.showAlert(
-              'Anchor Watch:',
-              'Server returned an error. This function may not be supported by your server.'
-            );
+            this.app.showAlert('Anchor Watch:', r.result.message);
           }
         }
       })
@@ -300,6 +298,13 @@ export class AppComponent {
       ? false
       : true;
     this.skres.setMapZoomRange();
+    this.app.saveConfig();
+  }
+
+  public invertAISTextColor() {
+    this.app.config.map.invertColor = this.app.config.map.invertColor
+      ? false
+      : true;
     this.app.saveConfig();
   }
 
@@ -356,7 +361,7 @@ export class AppComponent {
       if (this.selFavourite === -1) {
         this.selFavourite = 0;
       } else if (
-        this.selFavourite ==
+        this.selFavourite ===
         this.app.config.selections.pluginFavourites.length - 1
       ) {
         this.selFavourite = -1;
@@ -553,7 +558,7 @@ export class AppComponent {
       // use server trail data, keep minimal local trail data
       const lastseg = trailData.slice(-1);
       const lastpt =
-        lastseg.length != 0
+        lastseg.length !== 0
           ? lastseg[0].slice(-1)
           : trailData.length > 1
           ? trailData[trailData.length - 2].slice(-1)
@@ -593,7 +598,7 @@ export class AppComponent {
     if (e.action === 'save' && e.setting === 'config') {
       this.setDarkTheme(); // **  set theme **
       if (
-        this.app.data.trueMagChoice !=
+        this.app.data.trueMagChoice !==
         this.app.config.selections.headingAttribute
       ) {
         this.app.debug('True / Magnetic selection changed..');
@@ -1346,7 +1351,9 @@ export class AppComponent {
   // ** handle modify end event **
   public handleModifyEnd(e: {
     id: string;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     coords: any[];
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     coordsMetadata?: any[];
   }) {
     if (
@@ -1622,7 +1629,7 @@ export class AppComponent {
         this.setDarkTheme();
         if (
           e.result.self.resourceUpdates &&
-          e.result.self.resourceUpdates.length != 0
+          e.result.self.resourceUpdates.length !== 0
         ) {
           this.skres.processDelta(e.result.self.resourceUpdates);
         }