+++ src/app/modules/map/ol/lib/navigation/layer-arrival-circle.component.ts
@@ -80,13 +80,13 @@ export class ArrivalCircleComponent implements OnInit, OnDestroy, OnChanges {
       const properties: { [index: string]: any } = {};
 
       for (const key in changes) {
-        if (key == 'position' || key == 'radius') {
+        if (key === 'position' || key === 'radius') {
           this.parseValues();
           if (this.source) {
             this.source.clear();
             this.source.addFeatures(this.features);
           }
-        } else if (key == 'layerProperties') {
+        } else if (key === 'layerProperties') {
           this.layer.setProperties(properties, false);
         } else {
           properties[key] = changes[key].currentValue;