+++ src/app/modules/map/ol/lib/themes.ts
@@ -1,0 +1,13 @@
+// Map Themes
+
+export const LightTheme = {
+  labelText: {
+    color: '#333'
+  }
+};
+
+export const DarkTheme = {
+  labelText: {
+    color: 'yellow'
+  }
+};