+++ src/app/modules/skresources/notes/relatednotes-dialog.ts
@@ -2,7 +2,23 @@
  ********************************/
 
 import { Component, OnInit, Inject } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import {
+  MatDialogModule,
+  MatDialogRef,
+  MAT_DIALOG_DATA
+} from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatToolbarModule } from '@angular/material/toolbar';
+
+import { AddTargetPipe } from './safe.pipe';
+
 import { AppInfo } from 'src/app/app.info';
 
 /********* RelatedNotesDialog **********
@@ -11,7 +27,21 @@ import { AppInfo } from 'src/app/app.info';
     }
 ***********************************/
 @Component({
+  standalone: true,
   selector: 'ap-relatednotesdialog',
+  imports: [
+    CommonModule,
+    FormsModule,
+    MatDialogModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
+    MatTooltipModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatToolbarModule,
+    AddTargetPipe
+  ],
   templateUrl: `relatednotes-dialog.html`,
   styleUrls: ['notes.css']
 })